+++ src/doom/g_game.c
@@ -1824,9 +1824,6 @@ void G_LoadGame (char* name)
     M_StringCopy(savename, name, sizeof(savename));
     gameaction = ga_loadgame; 
 } 
- 
-#define VERSIONSIZE		16 
-
 
 int savedleveltime = 0; // [crispy] moved here for level time logging
 void G_DoLoadGame (void)